restructure code to allow function reusability
